+++ HELP.md
@@ -1,7 +1,7 @@
 # Read Me First
 The following was discovered as part of building this project:
 
-* The original package name 'com.liftoff.library-manager' is invalid and this project uses 'com.liftoff.librarymanager' instead.
+* The original package name 'com.liftoff.library-manager' is invalid and this project uses 'com.liftoff.libraryapp' instead.
 
 # Getting Started
