+++ src/main/java/com/liftoff/libraryapp/LibraryApp.java
@@ -1,13 +1,13 @@
-package com.liftoff.librarymanager;
+package com.liftoff.libraryapp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class LibraryManagerApplication {
+public class LibraryApp {
 
 	public static void main(String[] args) {
-		SpringApplication.run(LibraryManagerApplication.class, args);
+		SpringApplication.run(LibraryApp.class, args);
 	}
 
 }
