+++ src/main/java/com/liftoff/libraryapp/controllers/HelloController.java
@@ -1,10 +1,8 @@
-package com.liftoff.librarymanager.controllers;
+package com.liftoff.libraryapp.controllers;
 
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 public class HelloController {
