+++ src/main/java/com/liftoff/libraryapp/models/Book.java
@@ -1,10 +1,7 @@
-package com.liftoff.librarymanager.models;
+package com.liftoff.libraryapp.models;
 
 import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
 import java.util.Date;
-import java.util.Optional;
 
 @Entity
 public class Book {
