+++ src/test/java/com/liftoff/libraryapp/LibraryManagerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.liftoff.librarymanager;
+package com.liftoff.libraryapp;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
